pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sdcoh"
version = "0.2.0"
description = "Story Design Coherence — manage dependency graphs between story design documents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "ysttsu" }]
keywords = ["fiction", "writing", "coherence", "dependency-graph", "claude-code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openviking = ["httpx>=0.27"]
dev = ["pytest>=8.0", "pytest-tmp-files>=0.0.2"]

[project.scripts]
sdcoh = "sdcoh.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
